Addressing more feedback
