+++ src/Microsoft.CodeQuality.Analyzers/Core/Microsoft.CodeQuality.Analyzers.csproj
@@ -13,10 +13,5 @@
     <InternalsVisibleTo Include="Microsoft.CodeQuality.VisualBasic.Analyzers" />
     <InternalsVisibleTo Include="Microsoft.CodeQuality.Analyzers.UnitTests" />
   </ItemGroup>
-  <ItemGroup>
-    <EmbeddedResource Update="Security\MicrosoftSecurityAnalyzersResources.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-    </EmbeddedResource>
-  </ItemGroup>
   <Import Project="..\..\Utilities\Analyzer.Utilities.projitems" Label="Shared" />
 </Project>
