+++ src/Microsoft.NetFramework.Analyzers/Core/MicrosoftSecurityAnalyzersResources.resx
