+++ src/Microsoft.NetFramework.Analyzers/Core/xlf/MicrosoftSecurityAnalyzersResources.it.xlf
