+++ src/Microsoft.NetFramework.Analyzers/Core/xlf/MicrosoftSecurityAnalyzersResources.ja.xlf
