+++ src/Microsoft.NetFramework.Analyzers/Core/xlf/MicrosoftSecurityAnalyzersResources.ko.xlf
