+++ src/Microsoft.NetFramework.Analyzers/Core/xlf/MicrosoftSecurityAnalyzersResources.tr.xlf
